requirements.txt
discord.py
python-dotenv
requests
sqlalchemy
ruamel.yaml
schedule